오류해결
오류해결

+++ src/main/java/com/project/api/entity/MemberEntity.java
@@ -18,7 +18,7 @@ public class MemberEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer memberNo;
+    private Long memberNo;
 
     private String id;
     private String password;
